Created MeasurementEventFactory and MeasurementEvent classes.
The MeasurementEvent class provides an abstraction that contains all the information needed for the Injector class to inject javascript into the page to track an event for a plugin. The MeasurementEventFactory provides a function that the Injector class can use to produce a list of MeasurementEvent objects for a certain plugin.