Created MeasurementEventList class and WoocommerceEventList class.
MeasurementEventList is a template class that holds all the information needed to create a list of MeasurementEvent objects. It can be extended by a plugin-specific list to add its customized measurement event information. The WoocommerceEventList is an example of such a class and adds its woocommerce-specific information to create a list of trackable events.