Implemented Javascript injection to add event listeners to the Woocommerce events and log when one of those events occurs.
At this point, the actual css selectors for woocommerce need to be defined more clearly to make sure every event works for every theme. Another thing that is needed is to add javascript that checks if the element is actually on the page before we add an event listener to prevent the console from being overloaded with errors. On events that cause the page to reload, we are unable to see if the event fired so we need to come up with a solution for that as well. Lastly, unit tests still need to be implemented.