Cleaned up some Woocommerce selectors and found a solution to some Woocommerce elements not being on the page at the page's first load.
Before this commit we were not selecting some Woocommerce elements correctly because we were using the sibling operator (~) when it was not needed. Instead, I changed the body selector to be more specific so that we are sure the first selector is selecting the page's body element. The other problem was some of Woocommerce's elements were not on the page immediately on page load, but were added some time in the future after an Ajax request. On most of these requests, Woocommerce triggers actions on the page's body element, so by hooking into those events I was able to add an event listener after the page was loaded. Still need to implement the EventList classes for the rest of the plugins and add unit tests.