Created WPFormsList class
WPFormsList contains all the necessary info to support the WPForms plugin. Also changed the way javascript is injected into the page in the MeasurementEvent class to be more readable. Also used a for loop to iterate through each node in the javascript query which solves two problems. The first is now we don't try to add event listeners to elements that aren't on the current page. The other is that if there are multiple elements that we want to track for the same event e.g. form-submit, we now add event listeners to each element.