Updated measurement event list instantiation architecture.
Previously, each subclass of the MeasurementEventList class would store all the event data and then use that data to create a list of MeasurementEvent objects. That method of creation was redundant because that information is already stored in each event, and it also creates a highly coupled relationship between the MeasurementEventList class and the MeasurementEvent class. Now, each subclass directly creates MeasurementEvent objects and only stores those objects.