Cleaned up minor comments and code bits to be pull-request ready.
Two bugs remain to be fixed. The first has to do with Ninja Forms. We are able to track when a Ninja Form is submitted a single time, but not any times after that. The other bug has to do with Woocommerce. When the user is on the page where they are viewing the cart contents, we curently track three actions: checkout, remove_from_cart, and update_cart. Within Woocommerce, if one of the latter two actions fires, the document that contains those three elements is refreshed with an Ajax request, so our event listeners for all three elements is lost. We need to find a way to hook into that ajax request so we can reapply our event listeners when the ajax request occurs.