Now listening for all events on the root element.
Previously, we were attaching event listeners to specific elements. Now
we listen for all events with the root element, and during the capture
phase of event propogation so we will always capture it. This makes it
so that we don't need to worry if the elements are added dynamically
because we simply listen for all events on the root element. It also
fixes previous bugs that occurred when an event caused a separate element to be
reloaded dynamically so the existing event handlers were lost.